fix:week8 調整留言板 Session 內容；修正 transaction 內容並確保其能運作

+++ homeworks/week8/hw2/del_comments.php
@@ -1,28 +1,31 @@
 <?php
 	require_once('conn.php');
-	$tmpId = $_COOKIE["tmpId"];
-	$postId = $_POST['commentId'];
+	
+	$result = null;
 	$rtnCde = "success";
 
-	$stmt = $conn->prepare("SELECT c.id FROM marin_users_certificate a JOIN marin_users b ON a.username = b.username JOIN marin_comments c ON c.crt_user = b.username AND c.id = ?  WHERE a.id = ?");
-	$stmt->bind_param("is", $postId, $tmpId);
+	if(isset($_SESSION["userAccount"])){
+		$postId = $_POST['commentId'];
+		$stmt = $conn->prepare("SELECT c.id FROM marin_comments c WHERE c.crt_user = ? AND c.id = ?");
+		$stmt->bind_param("si", $_SESSION["userAccount"], $postId);
 
-	if($stmt->execute()){
-		$certificate = $stmt->get_result();
-		$stmt->close();
+		if($stmt->execute()){
+			$certificate = $stmt->get_result();
+			$stmt->close();
 
-		$certificateRow = $certificate->fetch_assoc();
-		$commentId = $certificateRow['id'];
-		$stmt = $conn->prepare("DELETE FROM marin_comments WHERE id = ? OR parent_id = ?");
-		$stmt->bind_param("ii", $commentId, $commentId);
-		if(!$stmt->execute()){
+			$certificateRow = $certificate->fetch_assoc();
+			$commentId = $certificateRow['id'];
+			$stmt = $conn->prepare("DELETE FROM marin_comments WHERE id = ? OR parent_id = ?");
+			$stmt->bind_param("ii", $commentId, $commentId);
+			if(!$stmt->execute()){
+				$rtnCde = "fail";
+			}
+			$stmt->close();
+		}
+		else{
+			$stmt->close();
 			$rtnCde = "fail";
 		}
-		$stmt->close();
-	}
-	else{
-		$stmt->close();
-		$rtnCde = "fail";
 	}
 	$conn->close();
